set as many tables as possible to InnoDb and utf8 in database
change the phpthumb.class.php in vendors/phpThumb to allow var $config_allow_src_above_docroot              = true;
because apparently doesnt work well with symlink required by capistrano

meio_duplicate is also changed conforming more to the mainsite way of doing things.
be ready to revise if it doesn't work.

clean_themed revised to remove ALL PRODUCT IMAGES except for default.jpg

+++ app/vendors/shells/clean_themed.php
@@ -45,13 +45,19 @@ class CleanThemedShell extends Shell {
                 $path = ROOT . DS . 'app' . DS . 'webroot' . DS. 'uploads' . DS . 'products' . DS;
        
                 $productImageFolder = new Folder($path);
-                $arrayOfProductImages = $productImageFolder->find('^default-.*');
+                //$arrayOfProductImages = $productImageFolder->find('^default-.*');
+                $arrayOfProductImages = $productImageFolder->find('.*');
                 
                 foreach($arrayOfProductImages as $fileName) {
-                        $file = new File($path . $fileName);
-                        $file->delete();
-                        $this->out($fileName . ' deleted');
+                        // we delete everything except default.jpg
+                        if ($fileName != 'default.jpg') {
+                                $file = new File($path . $fileName);
+                                $file->delete();
+                                $this->out($fileName . ' deleted');        
+                        }
                 }
+                
+                
                 
                 $this->out('now deleting the thumbs in APP');
                 $arrayOfThumbs = array('thumb',
@@ -65,14 +71,18 @@ class CleanThemedShell extends Shell {
                 foreach ($arrayOfThumbs as $folderName) {
                         $newPath = $path . $folderName;
                         $productImageFolder = new Folder($newPath);
-                        $arrayOfProductImages = $productImageFolder->find('^default-.*');
+                        //$arrayOfProductImages = $productImageFolder->find('^default-.*');
+                        $arrayOfProductImages = $productImageFolder->find('.*');
                         
                         $this->out('now deleting the thumbs in ' . $folderName);
                         
                         foreach($arrayOfProductImages as $fileName) {
-                                $file = new File($newPath . DS . $fileName);
-                                $file->delete();
-                                $this->out($fileName . ' deleted');
+                                // we delete everything except default.jpg
+                                if ($fileName != 'default.jpg') {
+                                        $file = new File($newPath . DS . $fileName);
+                                        $file->delete();
+                                        $this->out($fileName . ' deleted');
+                                }
                         }
                         
                 }
